fix remaining issues from squashing
