+++ src/components/relay-dashboard/Balance/components/SendForm/components/ResultScreen/ResultScreen.styles.ts
